+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - YYYY-MM-DD
+
+### Fixed
+
+- `pctx dev`
+  - Better error state reporting (e.g. invalid config, port already in use)
+  - Scroll out of bounds for tool details panel
+
 ## [v0.2.0] - 2025-11-19
 
 ### Added